requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
anyio==4.0.0
appdirs==1.4.4
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.6
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.4
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.7.0
bleach==6.0.0
blinker==1.6.2
Brotli==1.1.0
brotlipy==0.7.0
cachetools==5.3.1
cap-from-youtube==0.0.10
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
cmd2==2.4.3
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
cookiecutter==2.3.0
cryptography @ file:///croot/cryptography_1673298753778/work
cuda-python==12.2.0
cycler==0.11.0
Cython==3.0.0
datasets==2.14.5
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
dm-tree==0.1.8
docopt==0.6.2
docstring-to-markdown==0.12
docutils==0.20.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.2.1
fastjsonschema==2.18.0
filelock==3.12.2
flake8==6.0.0
flatbuffers==23.5.26
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.17.2
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
imread-from-url==0.1.3
inflection==0.5.1
install==1.3.5
intervaltree==3.1.0
ipykernel==6.25.2
ipython @ file:///croot/ipython_1674681422581/work
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
isort==5.12.0
jaraco.classes==3.3.0
jax==0.4.14
jaxlib==0.4.14+cuda11.cudnn86
jedi==0.18.2
jeepney==0.8.0
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter_client @ file:///croot/jupyter_client_1671703053786/work
jupyter_core==5.3.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.14.0
keyring==24.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.3
libclang==16.0.6
lit==16.0.6
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
MechanicalSoup==0.8.0
mistune==3.0.1
ml-dtypes==0.2.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mutagen==1.47.0
mxnet-cu110==2.0.0a0
MyNeuron==1.0.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.3
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu117==11.10.1.25
nvidia-cublas-cu12==12.2.4.5
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu117==11.7.50
nvidia-cuda-nvcc-cu11==2022.5.4
nvidia-cuda-nvcc-cu117==11.7.64
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.2.128
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu117==11.7.60
nvidia-cuda-runtime-cu12==12.2.128
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu116==8.4.0.27
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pyindex==1.0.9
nvidia-tensorrt==99.0.0
oauthlib==3.2.2
onnx==1.14.1
onnx-graphsurgeon==0.3.27
onnx-simplifier==0.4.33
onnx-tf==1.10.0
onnx2tf==1.15.18
onnxconverter-common==1.14.0
onnxruntime==1.12.1
onnxruntime-gpu==1.15.1
onnxsim==0.4.33
opencv-python==4.8.0.76
opt-einsum==3.3.0
optimum==1.13.1
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.11.2
patool==1.12
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.10.0
pluggy==1.3.0
ply==3.11
prettytable==0.7.2
progressbar2==3.34.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.18.0
pycuda==2022.2.2
pydeck==0.8.0
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.16.1
pylint==2.17.5
pylint-venv==3.0.2
pyls-spyder==0.4.0
Pympler==1.0.1
pyngrok==6.0.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
python-slugify==8.0.1
python-utils==3.5.2
pytoolconfig==1.2.5
pytools==2023.1.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rope==1.9.0
rpds-py==0.10.2
rsa==4.9
Rtree==1.0.1
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
SecretStorage==3.3.3
selenium==4.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six==1.16.0
smmap==5.0.0
sng4onnx==1.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
spyder==5.4.5
spyder-kernels==2.4.4
stack-data==0.6.2
stevedore==5.1.0
streamlit==1.27.2
sympy==1.12
tenacity==8.2.2
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-model-optimization==0.7.5
tensorflow-probability==0.21.0
tensorflow-serving-api==2.13.1
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
textdistance==4.5.0
tf2onnx==1.8.1
tflite-support==0.4.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
three-merge==0.1.1
tifffile==2023.7.18
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torch-summary==1.4.5
torch-tensorrt==1.4.0
torchaudio==2.0.1+cu118
torchsummary==1.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.2
trio==0.22.1
trio-websocket==0.10.3
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
ultralytics==8.0.158
uri-template==1.3.0
urllib3==1.26.13
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
whatthepatch==1.0.5
widgetsnbextension==4.0.8
wrapt==1.14.1
wsproto==1.2.0
wurlitzer==3.0.3
xxhash==3.3.0
yapf==0.40.1
yarg==0.1.9
yarl==1.9.2
yt-dlp==2023.7.6
zipp==3.17.0
